Fix recording in Safari by actually calling `stop`
That was a stupid oversight before and I'm not sure how it even works
in other browsers. This is quite the diff as I had to pull lots of
logic from the recording controls to the parent component.
